requirements.txt
django==1.11.3
psycopg2==2.7.3
ipython==6.0
flake8==3.3.0
flake8-mypy==17.3.3
django-countries==4.6.1
django-extensions==1.9.0
olefile==0.44
Pillow==4.3.0
django-role-permissions==2.2.0

